kubeadm-config.md: update the page to reflect current state
We have seen a number of questions around the usage of "config print".
Clarify the usage of this command. Add a note that the output should
not be taken literally.

Link to the "customizing components" page instead of the "kubelet
integration" page. The former already links to the later.

Remove note about kube-dns which hasn't been supported for a while.