kubeadm: update HA etcd guide for clarity and fix an issue
There were a couple of reported problems with this guide:

- The introductory paragraph talks about single control plane nodes
and does not mention the different options for HA etcd. Clear the
language to reduce the confusion and cross-link to the ha-topology
page.
- The hostname / IP detection in kubeadm can end
up with values not suitable for the certificates that kubeadm
generates for all etcd instances. Ensure that the hostnames / IPs
are pinned by the user in the example script.

Side cleanup related to the dockershim removal:
- Use containerd in the setup example and don't mention docker as
a requirement.