Fix links in security context page
The links to AppArmor etc are using redirection records. This is a bad
practice because when translating the page to a different language, the
link will become invalid. So we should use the actual target in the
English source.

This PR also fixes some incorrect syntax directives for code snippets.