--all-namespaces shorthand
As a beginner to kubectl, I must have typed (and re-typed due to typos) --all-namespaces many hundreds of times before I stumbled across the -A shorthand notation for this.  I sincerely would be grateful on behalf of all kubectl beginners if this very useful cheat could be added to this cheatsheet.  I have proposed in this PR a location quite near the beginning of the cheatsheet for this so that it is very hard to miss.