Improve node pressure eviction page
There are deprecated kubelet flags related to garbage collection. These flags were listed in the `kubelet-garbage-collection.md` page, which was removed by b1573ad31463dc25ec77cb1cbee7684ae944d7cd. Linking to the `kubelet` reference is not a good solution because users may still get confused what are the specific flags related to this behavior.

This PR restores the table with the flags updated according to 1.24 version of `kubelet --help`.